+++ src/extension.ts
@@ -26,6 +26,7 @@ export async function activate(context: vscode.ExtensionContext) {
     vscode.commands.registerCommand('voxpilot.toggleListening', () => engine?.toggle()),
     vscode.commands.registerCommand('voxpilot.pushToTalk', () => engine?.quickCapture()),
     vscode.commands.registerCommand('voxpilot.selectModel', () => engine?.selectModel()),
+    vscode.commands.registerCommand('voxpilot.selectAudioDevice', () => engine?.selectAudioDevice()),
     vscode.commands.registerCommand('voxpilot.sendToChat', () => engine?.sendLastToChat()),
     vscode.commands.registerCommand('voxpilot.clearCache', () => clearCache(context)),
     statusBar,